+++ src/main/java/com/mageddo/dnsproxyserver/dnsconfigurator/linux/ConfigureDPSHandler.java
@@ -1,23 +1,25 @@
-package com.mageddo.dnsproxyserver.dnsconfigurator.linux.systemdresolved;
+package com.mageddo.dnsproxyserver.dnsconfigurator.linux;
 
 import com.mageddo.conf.parser.Entry;
 import com.mageddo.conf.parser.Transformer;
 import com.mageddo.dnsproxyserver.dnsconfigurator.linux.DpsTokens;
+import com.mageddo.dnsproxyserver.dnsconfigurator.linux.EntryTypes;
 
 import java.util.Set;
+import java.util.function.Supplier;
 
 public class ConfigureDPSHandler implements Transformer {
 
-  private final String serverIP;
+  private final Supplier<String> dpsDnsLineBuilder;
 
-  public ConfigureDPSHandler(String serverIP) {
-    this.serverIP = serverIP;
+  public ConfigureDPSHandler(Supplier<String> dpsDnsLineBuilder) {
+    this.dpsDnsLineBuilder = dpsDnsLineBuilder;
   }
 
   @Override
   public String handle(Entry entry) {
     return switch (entry.getType().name()) {
-      case EntryTypes.DPS_SERVER -> buildDNSLine(this.serverIP);
+      case EntryTypes.DPS_SERVER -> this.dpsDnsLineBuilder.get();
       case EntryTypes.SERVER -> DpsTokens.comment(entry.getLine());
       default -> entry.getLine();
     };
@@ -26,12 +28,9 @@ public class ConfigureDPSHandler implements Transformer {
   @Override
   public String after(boolean fileHasContent, Set<String> foundEntryTypes) {
     if (!fileHasContent || !foundEntryTypes.contains(EntryTypes.DPS_SERVER)) {
-      return buildDNSLine(this.serverIP);
+      return this.dpsDnsLineBuilder.get();
     }
     return null;
   }
 
-  static String buildDNSLine(String serverIP) {
-    return "DNS=" + serverIP + " # dps-entry";
-  }
 }
