+++ src/test/java/com/mageddo/dnsproxyserver/dnsconfigurator/linux/systemdresolved/ResolvedConfiguratorTest.java
@@ -1,5 +1,6 @@
 package com.mageddo.dnsproxyserver.dnsconfigurator.linux.systemdresolved;
 
+import com.mageddo.dnsproxyserver.dnsconfigurator.linux.ResolvedConfigurator;
 import com.mageddo.dnsproxyserver.templates.IpTemplates;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
